bluetoothctl runs forever on some devices
Apparently bluetoothctl makes this scipt to hang. I was first using this addon on HAOS, where it ran without issues. Than I migrated to HA supervised, where it did not start, it just printed the output of the uname command then waited forever. With this proposed change the Bluetooth address will be printed only when Bluetooth is configured (vs Ethernet).