Add pvoSID example to sbfspot README
Because I've been struggling with not being able to upload anything to pvoutput.org - while I
even copy and pasted the System ID from the website to the configuration to not mix up
any numbers - I looked into the source code and found better instructions in the file
`sbfspot/rootfs/usr/bin/sbfspot/SBFspotUpload.default.cfg`. This is why I copy and pasted
them into the README, so that other people don't need to struggle to understand what
the numbers before the `:` in the example mean.